+++ src/meadowgrid/job_worker_main.py
@@ -5,7 +5,7 @@ import logging
 import multiprocessing
 import os
 import os.path
-from typing import Optional, ContextManager, Dict
+from typing import Iterator, Optional, Dict
 
 import meadowgrid.job_worker
 from meadowgrid.config import DEFAULT_COORDINATOR_HOST, DEFAULT_COORDINATOR_PORT
@@ -50,7 +50,7 @@ def main_in_child_process(
     available_resources: Optional[Dict[str, float]] = None,
     coordinator_host: Optional[str] = None,
     coordinator_port: Optional[int] = None,
-) -> ContextManager[None]:
+) -> Iterator[Optional[int]]:
     """
     Launch worker in a child process. Usually for unit tests. For debugging, it's better
     to just run job_worker_main.py manually as a standalone process so you can debug it,
@@ -64,10 +64,16 @@ def main_in_child_process(
     server_process.start()
 
     try:
+        logging.info(f"Process started. Pid: {server_process.pid}")
         yield server_process.pid
-        server_process.terminate()
     finally:
-        server_process.kill()
+        server_process.terminate()
+        logging.info("Process terminated. Waiting up to 5 seconds for exit...")
+        server_process.join(5)
+        logging.info(f"Process exited with code {server_process.exitcode}")
+        if server_process.is_alive():
+            logging.info("Process alive after termination, killing.")
+            server_process.kill()
 
 
 def command_line_main():
